Add tabs to worklist, move WorkList to WorkListTabContent.
